Move all code from OrderCreationFormViewModel to OrderCreationViewModel
